requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
bitarray==3.0.0
cached-property==2.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
ckzg==2.0.1
cytoolz==1.0.1
eth-account==0.11.2
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.2.0
frozenlist==1.5.0
hexbytes==0.3.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lru-dict==1.2.0
multidict==6.1.0
packaging==24.2
parsimonious==0.10.0
propcache==0.2.1
protobuf==5.29.3
py-ecc==7.0.1
py-evm==0.10.1b2
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pyunormalize==16.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
rpds-py==0.22.3
safe-eth-py==6.4.0
safe-pysha3==1.0.4
sortedcontainers==2.4.0
toolz==1.0.0
trie==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
web3==6.20.2
websockets==14.2
yarl==1.18.3
